requirements.txt
MySQL-python==1.2.5
requests==2.9.1
pytz==2016.3
Django==1.8.4
djangorestframework
django-filter
django-cors-headers==1.1.0
gunicorn==19.1.1
ipython